Added components to Main
